+++ pypeman/default_settings.py
@@ -71,9 +71,14 @@ DISABLED_PLUGINS = []
 LOGGING = {
     "version": 1,
     "disable_existing_loggers": False,
+    "filters": {
+        "context": {
+            "()": "pypeman.helpers.logging.LogContextFilter",
+        },
+    },
     "formatters": {
         "verbose": {
-            "format": "%(levelname)s %(asctime)s %(name)s %(module)s %(message)s",
+            "format": "%(levelname)s %(asctime)s %(name)s %(channel)s%(msg_id)s%(message)s",
         },
     },
     "handlers": {
@@ -81,6 +86,7 @@ LOGGING = {
             "level": "DEBUG",
             "class": "logging.StreamHandler",
             "formatter": "verbose",
+            "filters": ["context"],
         },
     },
     "loggers": {